allow user to decide to install virtualenv or proceed with a different installation type

+++ setup.sh
@@ -16,6 +16,11 @@ function default(){
     echo "$__resultvar"
 }
 
+function invalid(){
+    echo "Invalid selection, exiting setup program"
+    exit 1
+}
+
 home=~
 eval home=$home/passdb
 echo "If not using defaults for the following paths please use full filepath"
@@ -56,29 +61,51 @@ case "$pinstall" in
     python -m pip install -r requirements.txt --user
     ;;
 "2")
-    read -rp "What would you like to call the venv? (Default pkpass): " venv
-    if [ -z "$venv" ];then
-        venv="pkpass"
-    fi
-    python -c "import virtualenv"
+    #virtualenv install
+    #check to see if virtualenv package exists return 0 is true
+    python -c "import virtualenv" 2>/dev/null
+
     if [ "$?" == "1" ]; then
-        read -rp "virtualenv package not detected would you like to install as root(0) or user(1)?" vinstall
-        if [[ "$vinstall" == "0" ]]; then
-            sudo python -m pip install virtualenv
-        elif [[ "$vinstall" == "1" ]]; then
-            python -m pip install virtualenv --user
+        read -rp "virtualenv package not dected would you like to install it?(y/n)" installvenv
+        #make user input lower case
+        installvenv="$(echo "$installvenv" | tr '[:upper:]' '[:lower:]')"
+
+        if [[ "$installvenv" == "y" ]]; then
+            read -rp "Would you like to install virtualenv as root(0) or user(1)?" vinstall
+            if [[ "$vinstall" == "0" ]]; then
+                sudo python -m pip install virtualenv
+            elif [[ "$vinstall" == "1" ]]; then
+                python -m pip install virtualenv --user
+            else
+                invalid
+            fi
+
+        elif [[ "$installvenv" == "n" ]]; then
+            read -rp "how would you like to install the python requirements instead? root(0)/user(1): " preq
+            if [[ "$preq" == 0 ]]; then
+                sudo python -m pip install -r requirements.txt
+            elif [[ "$preq" == 1 ]]; then
+                python -m pip install -r requirements.txt --user
+            fi
+
         else
-            echo "Invalid selection, exiting program"
-            exit 1
+            invalid
         fi
     fi
-    python -m virtualenv "$venv"
-    source "$venv"/bin/activate
-    pip install -r requirements.txt
+    #don't process virtualenv install if user says no otherwise do it
+    if [[ "$installvenv" != "n" ]]; then
+        read -rp "What would you like to call the venv? (Default pkpass): " venv
+        if [ -z "$venv" ];then
+            venv="pkpass"
+        fi
+
+        python -m virtualenv "$venv"
+        source "$venv"/bin/activate
+        pip install -r requirements.txt
+    fi
     ;;
 *)
-    echo "Invalid selection, exiting program"
-    exit 1
+    invalid
 esac
 
 echo "testing versions of openssl and pkcs15-tool"
